+++ src/main/java/vn/framgia/dao/impl/UserDAOImpl.java
@@ -8,7 +8,7 @@ import vn.framgia.model.User;
  * Created by FRAMGIA\duong.van.tien on 06/03/2017.
  *
  */
-public class UserDAOImpl extends GenericDAO<User, Long> implements IUserDAO {
+public class UserDAOImpl extends GenericDAO<User, Integer> implements IUserDAO {
     public UserDAOImpl() {
         super.setPersistentClass(User.class);
     }
